Recode cluster scripts to use Rstan instead of cmdstan
Swtiching to Rstan instead of cmdstan because cmdstan models have been havning on the cluster. Rstan appears to yield better results. This may be due to the version of cmdstan I've been using on the cluster being outdated. Unfortunately, the version of gcc needed to complile the newer versions of cmdstan has not yet been installed on the cluster, and compilation fails with the older versions. Rstan is the next best option.